Fix a segfault
Before attempting to read a message in on_read(), check whether the
connection has been closed. This makes the plugin somewhat usable again.

For debugging I often use murmur (the server) and mumble (the client).
Apparently there are now problems with mumble related to audio settings.
When selecting the ALSA driver mumble uses about 20 % cpu when idle and
muted and deafened, while cpu usage should be around 0. When selecting
the null driver, which I suppose should disable any audio processing,
mumble takes 100 % cpu.

Another thing is that when mumble is receiving a message from murmur, it
completely freezes for a couple of seconds. Running with strace the last
syscall before the freeze is futex(). If ran as root this freezing
behavior goes away, but the high cpu usage still remains.

I'm guessing this is related to me recently removing PulseAudio, because it
is broken by design for all of my use cases. The problem with pulseaudio in
short: if audio is played locally, the pulseaudio client library should use
ALSA directly instead of the client-server indirection, which is
unnecessarily eating cpu.

PulseAudio seems to be the default "audio driver" on most linux distros
these days. I guess this is because it works for both use cases: local and
remote playback. On the other hand, it completely fails for the most common
use case: local playback. Then there is central software like Firefox that
don't seem to support more than one built-in audio driver, which then has to
be PulseAudio. There is a hack called apulse with which it is possible to
run Firefox-with-pulseaudio-driver without running the pulseaudio server.
Related bug report: https://bugzilla.mozilla.org/show_bug.cgi?id=1345661
Entertaining. :)

Unfortunately apulse doesn't seem to work with mumble. So another setback.
Maybe related: https://github.com/mumble-voip/mumble/issues/4171

Me feels like moving to a static compiled distro, where after a new package
is installed, not every other program is in danger to get broken.