Remove i18n
There is not much to translate anyway.
This can be reconsidered later.

Continuation to previous commit message, about mumble freezing:
This happened in dbus notification code, probably because dbus was
incorrectly configured, or something. D-Bus is more likely to work properly
on a full desktop environment like KDE or GNOME or XFCE, I guess. But the
issue can also be fixed by compiling mumble without dbus support (or by
disabling all notifications).

About high cpu usage:
It is the ALSA audio input thread that uses cpu even if mumble is muted, or
capture is disabled on kernel level. It is possible to stop this thread by
finding and selecting an audio input driver that "cannot be opened". Fair
enough.